refactor: modularize monolithic useStrategyBot hook and worker backend
