+++ app/client/src/hooks/useStrategyBot.ts
@@ -1,19 +1,14 @@
 import type {
   Candle,
-  OptionData,
-  IndicatorsResult,
-  V3OrderType,
   AllSignalData,
-  FinalSignal,
   ActivePosition,
   PositionLeg,
   ExecutionMode,
-  PaperAccountSummary,
-  PaperTrade,
   VrdData,
-  NiftySentiment,
-  UpstoxNewsItem,
-  NewsAlert,
+  IndicatorsResult,
+  FinalSignal,
+  PaperTrade,
+  PaperAccountSummary,
 } from '@/lib/types'
 import {
   useState,
@@ -23,11 +18,6 @@ import {
   useCallback,
 } from 'react'
 import { computeAllIndicators, getOtmStrike } from '@/lib/indicators'
-import {
-  evaluateGlobalSentiment,
-  evaluatePCR,
-  getV3Signal,
-} from '@/lib/v3Sentiment'
 import { notify } from '@/lib/notifications'
 import {
   runHardStopChecks,
@@ -36,28 +26,25 @@ import {
 } from '@/lib/strategyEngine'
 import { getStrategyConfig } from '@/lib/strategyConfig'
 import { fetchPaperAccount } from '@/lib/paperTrading'
-import { scoreStraddleIV, classifyNews } from '@/lib/vrdSignals'
+import { scoreStraddleIV } from '@/lib/vrdSignals'
 import { appendTick } from '@/lib/tickLog'
+import {
+  isStaticIpRestrictionError,
+  isPaperPosition,
+  getLotSizeForSymbol,
+} from '@/lib/syntheticCalculators'
+import type { SourceStatus, BotLog, GlobalIndexItem } from '@/lib/marketService'
+import {
+  mkLog,
+  safeFetch,
+  fetchMarket,
+  fetchMarketSentiment,
+} from '@/lib/marketService'
+
+export type { SourceStatus, BotLog, GlobalIndexItem }
 
 // ─── Types ─────────────────────────────────────────────────────────────────────
 export type BotState = 'IDLE' | 'RUNNING' | 'ORDERED' | 'STOPPED'
-export type SourceStatus = 'ok' | 'error' | 'stale' | 'pending' | 'unknown'
-
-export interface BotLog {
-  id: string
-  ts: string
-  level: 'info' | 'warn' | 'error' | 'debug'
-  source: string
-  msg: string
-}
-
-export interface GlobalIndexItem {
-  symbol: string
-  last_price: number
-  change_per: number
-  net_change?: number
-  [key: string]: unknown
-}
 
 export interface BotStatus {
   state: BotState
@@ -89,7 +76,6 @@ const KEYS = {
   vrdCache: 'algo-trade:vrd-cache', // { data: VrdData; savedAt: string }
   logs: 'algo-trade:bot-logs', // BotLog[] (last 200)
   snapshot: 'algo-trade:bot-snapshot',
-  proxyHistory: 'algo-trade:proxy-history',
 }
 const MAX_LOGS = 200
 const VRD_CACHE_MAX_MS = 30 * 60 * 1000 // 30 minutes
@@ -105,69 +91,6 @@ type BotSnapshot = Pick<
   | 'sourceStatus'
   | 'globalIndices'
 >
-
-// ─── Log factory ───────────────────────────────────────────────────────────────
-function mkLog(level: BotLog['level'], source: string, msg: string): BotLog {
-  return {
-    id: `${Date.now()}-${Math.random().toString(36).slice(2, 6)}`,
-    ts: new Date().toISOString(),
-    level,
-    source,
-    msg,
-  }
-}
-
-// ─── Safe JSON fetch — returns [data, null] or [null, errorMsg] ───────────────
-async function safeFetch<T>(
-  input: RequestInfo,
-  init?: RequestInit,
-): Promise<[T | null, string | null]> {
-  try {
-    const res = await fetch(input, init)
-    const data = (await res.json()) as
-      | T
-      | {
-          error?: string
-          message?: string
-          errors?: {
-            message?: string
-            errorCode?: string
-            error_code?: string
-          }[]
-        }
-    if (!res.ok) {
-      const errData = data as {
-        error?: string
-        message?: string
-        errors?: { message?: string; errorCode?: string; error_code?: string }[]
-      }
-      const detail =
-        errData.error ??
-        errData.message ??
-        errData.errors
-          ?.map((error) => error.message ?? error.errorCode ?? error.error_code)
-          .filter(Boolean)
-          .join(', ')
-      return [
-        null,
-        detail
-          ? `HTTP ${res.status} ${res.statusText}: ${detail}`
-          : `HTTP ${res.status} ${res.statusText}`,
-      ]
-    }
-    if (
-      data &&
-      typeof data === 'object' &&
-      'error' in data &&
-      (data as { error?: string }).error
-    ) {
-      return [null, String((data as { error: string }).error)]
-    }
-    return [data as T, null]
-  } catch (e) {
-    return [null, (e as Error).message]
-  }
-}
 
 // ─── Persist & load logs ───────────────────────────────────────────────────────
 function loadLogs(): BotLog[] {
@@ -231,44 +154,6 @@ function loadSnapshot(): Partial<BotSnapshot> {
   }
 }
 
-function loadProxyHistory(): { date: string; netPosition: number }[] {
-  try {
-    return JSON.parse(localStorage.getItem(KEYS.proxyHistory) ?? '[]') as {
-      date: string
-      netPosition: number
-    }[]
-  } catch {
-    return []
-  }
-}
-
-function saveProxyHistory(history: { date: string; netPosition: number }[]) {
-  try {
-    localStorage.setItem(KEYS.proxyHistory, JSON.stringify(history.slice(-20)))
-  } catch {
-    /* ignore */
-  }
-}
-
-function updateProxyHistory(netPosition: number): number | null {
-  const today = new Date().toISOString().split('T')[0]
-  const history = loadProxyHistory().filter((item) => item.date)
-  const withoutToday = history.filter((item) => item.date !== today)
-  const nextHistory = [...withoutToday, { date: today, netPosition }].sort(
-    (left, right) => right.date.localeCompare(left.date),
-  )
-  saveProxyHistory(nextHistory)
-
-  if (netPosition >= 0) return null
-
-  let consecutiveShortDays = 0
-  for (const item of nextHistory) {
-    if (item.netPosition < 0) consecutiveShortDays += 1
-    else break
-  }
-  return consecutiveShortDays || null
-}
-
 // ─── Load persisted bot state ─────────────────────────────────────────────────
 function loadPersisted(): Partial<BotStatus> {
   try {
@@ -318,931 +203,6 @@ function loadPersisted(): Partial<BotStatus> {
       logs: [],
     }
   }
-}
-
-function clamp(value: number, min: number, max: number): number {
-  return Math.min(max, Math.max(min, value))
-}
-
-function evaluateNiftySentimentFromAdvanceCount(
-  advances: number | null,
-): NiftySentiment {
-  if (advances === null || Number.isNaN(advances)) return 'neutral'
-  if (advances >= 39) return 'very bullish'
-  if (advances >= 29) return 'bullish'
-  if (advances >= 23) return 'neutral'
-  if (advances >= 13) return 'bearish'
-  return 'very bearish'
-}
-
-function isStaticIpRestrictionError(
-  message: string | null | undefined,
-): boolean {
-  if (!message) return false
-  const normalized = message.toLowerCase()
-  return (
-    normalized.includes('static ip restrictions') ||
-    normalized.includes('no static ip has been configured')
-  )
-}
-
-function isPaperPosition(position: ActivePosition | null): boolean {
-  return position?.executionMode === 'paper'
-}
-
-// ─── Synthetic MMI from Upstox data ──────────────────────────────────────────
-function computeMMI(
-  vix: number | null,
-  rsiValue: number,
-  pcrValue: number,
-): { score: number; label: string } {
-  const vixScore =
-    vix === null
-      ? 50
-      : vix > 25
-        ? 10
-        : vix > 20
-          ? 25
-          : vix > 16
-            ? 40
-            : vix > 13
-              ? 55
-              : vix > 10
-                ? 70
-                : 80
-
-  const rsiScore =
-    rsiValue < 30
-      ? 15
-      : rsiValue < 40
-        ? 30
-        : rsiValue < 50
-          ? 45
-          : rsiValue < 60
-            ? 55
-            : rsiValue < 70
-              ? 65
-              : 80
-
-  const pcrScore =
-    pcrValue > 1.5
-      ? 20
-      : pcrValue > 1.0
-        ? 35
-        : pcrValue > 0.8
-          ? 55
-          : pcrValue > 0.6
-            ? 65
-            : 80
-
-  const score = Math.round(vixScore * 0.4 + rsiScore * 0.3 + pcrScore * 0.3)
-
-  const label =
-    score < 25
-      ? 'Extreme Fear'
-      : score < 40
-        ? 'Fear'
-        : score < 55
-          ? 'Neutral'
-          : score < 70
-            ? 'Greed'
-            : 'Extreme Greed'
-
-  return { score, label }
-}
-
-// ─── Straddle IV from ATM option chain Greeks ─────────────────────────────────
-function computeStraddleIV(
-  optionChain: import('@/lib/types').OptionData[],
-  niftyLtp: number,
-  vix: number | null,
-): {
-  currentIv: number | null
-  averageIv: number | null
-  percentAboveAvg: number | null
-} {
-  if (!optionChain.length || niftyLtp === 0)
-    return { currentIv: null, averageIv: null, percentAboveAvg: null }
-  const atm = optionChain.reduce((prev, curr) =>
-    Math.abs(curr.strike_price - niftyLtp) <
-    Math.abs(prev.strike_price - niftyLtp)
-      ? curr
-      : prev,
-  )
-  const callIv = atm.call_options.option_greeks?.iv ?? null
-  const putIv = atm.put_options.option_greeks?.iv ?? null
-  if (callIv === null || putIv === null)
-    return { currentIv: null, averageIv: vix, percentAboveAvg: null }
-  const avgStrikeIv = (callIv + putIv) / 2
-  // Compare ATM straddle IV to VIX (VIX IS the 30-day expected IV benchmark)
-  const pct =
-    vix && vix > 0
-      ? parseFloat((((avgStrikeIv - vix) / vix) * 100).toFixed(1))
-      : null
-  return {
-    currentIv: parseFloat(avgStrikeIv.toFixed(2)),
-    averageIv: vix,
-    percentAboveAvg: pct,
-  }
-}
-
-function getAtmWindow(
-  optionChain: OptionData[],
-  niftyLtp: number,
-  windowSize = 5,
-): OptionData[] {
-  return [...optionChain]
-    .sort(
-      (left, right) =>
-        Math.abs(left.strike_price - niftyLtp) -
-        Math.abs(right.strike_price - niftyLtp),
-    )
-    .slice(0, windowSize)
-}
-
-function computeProxyFlow(optionChain: OptionData[], niftyLtp: number) {
-  if (!optionChain.length || niftyLtp === 0) {
-    return {
-      longPct: null as number | null,
-      shortPct: null as number | null,
-      netPosition: null as number | null,
-      consecutiveShortDays: null as number | null,
-    }
-  }
-
-  const window = getAtmWindow(optionChain, niftyLtp)
-  const callOi = window.reduce(
-    (sum, item) => sum + (item.call_options.market_data.oi ?? 0),
-    0,
-  )
-  const putOi = window.reduce(
-    (sum, item) => sum + (item.put_options.market_data.oi ?? 0),
-    0,
-  )
-  const totalOi = callOi + putOi
-  if (totalOi === 0) {
-    return {
-      longPct: null,
-      shortPct: null,
-      netPosition: null,
-      consecutiveShortDays: null,
-    }
-  }
-
-  const putShare = (putOi / totalOi) * 100
-  const longPct = parseFloat(clamp(putShare, 5, 95).toFixed(1))
-  const shortPct = parseFloat((100 - longPct).toFixed(1))
-  const netPosition = putOi - callOi
-  const consecutiveShortDays = updateProxyHistory(netPosition)
-
-  return {
-    longPct,
-    shortPct,
-    netPosition,
-    consecutiveShortDays,
-  }
-}
-
-function computeProxyValuation(
-  _niftyLtp: number,
-  indicators: IndicatorsResult,
-  vix: number | null,
-  adRatio: number | null,
-): { pe: number; label: string } {
-  const rsiStretch = (indicators.rsi.value - 50) / 8
-  const trendStretch =
-    indicators.bollinger.trend === 'Up'
-      ? 1.25
-      : indicators.bollinger.trend === 'Down'
-        ? -1.25
-        : 0
-  const volStretch = vix === null ? 0 : (15 - vix) / 5
-  const breadthStretch =
-    adRatio === null ? 0 : adRatio > 1.4 ? -0.75 : adRatio < 0.8 ? 1 : 0
-  const premiumStretch =
-    indicators.pcrValue > 1.1 ? -0.75 : indicators.pcrValue < 0.8 ? 1.25 : 0
-  const base =
-    21 +
-    rsiStretch +
-    trendStretch +
-    volStretch +
-    breadthStretch +
-    premiumStretch
-  const pe = parseFloat(clamp(base, 16, 30).toFixed(1))
-  const label =
-    pe < 18
-      ? 'Synthetic undervaluation'
-      : pe > 24
-        ? 'Synthetic overvaluation'
-        : 'Synthetic fair value'
-  return { pe, label }
-}
-
-// ─── Market Sentiment data (replaces VRD fetch) ───────────────────────────────
-async function fetchMarketSentiment(
-  token: string,
-  addLog: (l: BotLog) => void,
-  sourceUpdate: (k: string, s: SourceStatus) => void,
-  optionChain: OptionData[],
-  indicators: IndicatorsResult,
-  breadth: {
-    advances: number
-    declines: number
-    ratio: number
-    total: number
-  } | null,
-  giftNifty: VrdData['giftNifty'],
-): Promise<VrdData> {
-  sourceUpdate('vix', 'pending')
-  sourceUpdate('upstox/fii', 'pending')
-  sourceUpdate('upstox/dii', 'pending')
-  sourceUpdate('upstox/pcr', 'pending')
-  sourceUpdate('upstox/max-pain', 'pending')
-  sourceUpdate('synthetic/value', 'pending')
-  sourceUpdate('upstox/news', 'pending')
-
-  const latestExpiry = optionChain[0]?.expiry ?? ''
-
-  const [vixRes, fiiRes, diiRes, pcrRes, maxPainRes, newsRes] =
-    await Promise.allSettled([
-      safeFetch<{ vix: number | null }>('/api/market/vix', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      }),
-      safeFetch<{
-        status: string
-        data?: Record<
-          string,
-          {
-            time_stamp: number
-            total_long_contracts: number
-            total_short_contracts: number
-            buy_amount: number
-            sell_amount: number
-          }[]
-        >
-      }>('/api/market/upstox/fii', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      }),
-      safeFetch<{
-        status: string
-        data?: Record<
-          string,
-          {
-            time_stamp: number
-            buy_amount: number
-            sell_amount: number
-          }[]
-        >
-      }>('/api/market/upstox/dii', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      }),
-      safeFetch<{
-        value: number | null
-      }>('/api/market/upstox/pcr', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token, expiry: latestExpiry }),
-      }),
-      safeFetch<{
-        status: string
-        data?: {
-          max_pain: number
-        }
-      }>('/api/market/upstox/max-pain', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token, expiry: latestExpiry }),
-      }),
-      safeFetch<{
-        status: string
-        data?: Record<string, UpstoxNewsItem[]>
-      }>('/api/market/upstox/news', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          token,
-          category: 'instrument_keys',
-          instrumentKeys: 'NSE_INDEX|Nifty 50',
-        }),
-      }),
-    ])
-
-  // VIX
-  let vix: number | null = null
-  if (vixRes.status === 'fulfilled' && !vixRes.value[1]) {
-    vix = vixRes.value[0]?.vix ?? null
-    addLog(
-      mkLog(
-        vix !== null ? 'info' : 'warn',
-        'vix',
-        vix !== null ? 'VIX=' + vix : 'VIX returned null',
-      ),
-    )
-    sourceUpdate('vix', vix !== null ? 'ok' : 'error')
-  } else {
-    const err =
-      vixRes.status === 'fulfilled'
-        ? (vixRes.value[1] ?? 'unknown')
-        : 'fetch failed'
-    addLog(mkLog('error', 'vix', err))
-    sourceUpdate('vix', 'error')
-  }
-
-  // FII
-  let fiiLongShort: {
-    longPct: number | null
-    shortPct: number | null
-    shortPctTrend: 'Rising' | 'Falling' | 'Stable' | null
-  } | null = null
-  let fiiPositioning: {
-    netPosition: number | null
-    consecutiveShortDays: number | null
-  } | null = null
-
-  if (
-    fiiRes.status === 'fulfilled' &&
-    !fiiRes.value[1] &&
-    fiiRes.value[0]?.data
-  ) {
-    const fiiData = fiiRes.value[0].data
-    const indexFutures = fiiData['NSE_FO|INDEX_FUTURES'] ?? []
-    const sortedFii = [...indexFutures].sort(
-      (a, b) => b.time_stamp - a.time_stamp,
-    )
-    const latestFii = sortedFii[0]
-
-    if (latestFii) {
-      const long = latestFii.total_long_contracts ?? 0
-      const short = latestFii.total_short_contracts ?? 0
-      const total = long + short
-
-      if (total > 0) {
-        let shortPctTrend: 'Rising' | 'Falling' | 'Stable' | null = null
-        if (sortedFii.length >= 2) {
-          const getShortPct = (entry: (typeof sortedFii)[0]) => {
-            const l = entry.total_long_contracts ?? 0
-            const s = entry.total_short_contracts ?? 0
-            return l + s > 0 ? (s / (l + s)) * 100 : 0
-          }
-          const todayShortPct = getShortPct(sortedFii[0])
-          const comparisonEntry = sortedFii[Math.min(sortedFii.length - 1, 3)]
-          const pastShortPct = getShortPct(comparisonEntry)
-
-          if (todayShortPct - pastShortPct > 1.5) {
-            shortPctTrend = 'Rising'
-          } else if (pastShortPct - todayShortPct > 1.5) {
-            shortPctTrend = 'Falling'
-          } else {
-            shortPctTrend = 'Stable'
-          }
-        }
-
-        fiiLongShort = {
-          longPct: parseFloat(((long / total) * 100).toFixed(1)),
-          shortPct: parseFloat(((short / total) * 100).toFixed(1)),
-          shortPctTrend,
-        }
-        fiiPositioning = {
-          netPosition: long - short,
-          consecutiveShortDays: 0,
-        }
-
-        let shortDays = 0
-        for (const entry of sortedFii) {
-          const entryNet =
-            (entry.total_long_contracts ?? 0) -
-            (entry.total_short_contracts ?? 0)
-          if (entryNet < 0) {
-            shortDays++
-          } else {
-            break
-          }
-        }
-        fiiPositioning.consecutiveShortDays = shortDays || null
-
-        addLog(
-          mkLog(
-            'info',
-            'fii',
-            `FII Futures: L=${fiiLongShort.longPct}% S=${fiiLongShort.shortPct}% (${fiiLongShort.shortPctTrend}) Net=${fiiPositioning.netPosition}`,
-          ),
-        )
-        sourceUpdate('upstox/fii', 'ok')
-      } else {
-        sourceUpdate('upstox/fii', 'error')
-      }
-    } else {
-      sourceUpdate('upstox/fii', 'error')
-    }
-  } else {
-    sourceUpdate('upstox/fii', 'error')
-  }
-
-  // Fallback to proxy/synthetic FII if Upstox FII failed
-  const niftyLtp = optionChain[0]?.underlying_spot_price ?? 0
-  if (!fiiLongShort || !fiiPositioning) {
-    const proxyFlow = computeProxyFlow(optionChain, niftyLtp)
-    fiiLongShort =
-      proxyFlow.longPct !== null && proxyFlow.shortPct !== null
-        ? {
-            longPct: proxyFlow.longPct,
-            shortPct: proxyFlow.shortPct,
-            shortPctTrend: 'Stable',
-          }
-        : null
-    fiiPositioning =
-      proxyFlow.netPosition !== null
-        ? {
-            netPosition: proxyFlow.netPosition,
-            consecutiveShortDays: proxyFlow.consecutiveShortDays,
-          }
-        : null
-    addLog(
-      mkLog(
-        'warn',
-        'fii',
-        'FII API unavailable; fell back to synthetic flow options estimate',
-      ),
-    )
-    sourceUpdate('upstox/fii', 'stale')
-  }
-
-  // DII
-  if (
-    diiRes.status === 'fulfilled' &&
-    !diiRes.value[1] &&
-    diiRes.value[0]?.data
-  ) {
-    const diiData = diiRes.value[0].data
-    const cashList = diiData['NSE_EQ|CASH'] ?? []
-    const latestDii = [...cashList].sort(
-      (a, b) => b.time_stamp - a.time_stamp,
-    )[0]
-    if (latestDii) {
-      const netCash = (latestDii.buy_amount ?? 0) - (latestDii.sell_amount ?? 0)
-      addLog(
-        mkLog(
-          'info',
-          'dii',
-          `DII Cash Net: ${(netCash / 10000000).toFixed(2)} Cr`,
-        ),
-      )
-      sourceUpdate('upstox/dii', 'ok')
-    } else {
-      sourceUpdate('upstox/dii', 'error')
-    }
-  } else {
-    sourceUpdate('upstox/dii', 'error')
-  }
-
-  // PCR
-  let officialPcr: number | null = null
-  if (pcrRes.status === 'fulfilled' && !pcrRes.value[1]) {
-    officialPcr = pcrRes.value[0]?.value ?? null
-    if (officialPcr !== null) {
-      addLog(
-        mkLog('info', 'upstox/pcr', `Upstox PCR=${officialPcr.toFixed(3)}`),
-      )
-      sourceUpdate('upstox/pcr', 'ok')
-    } else {
-      sourceUpdate('upstox/pcr', 'error')
-    }
-  } else {
-    sourceUpdate('upstox/pcr', 'error')
-  }
-
-  const effectivePcr = officialPcr ?? indicators.pcrValue
-  addLog(mkLog('info', 'pcr', 'Option PCR=' + effectivePcr.toFixed(3)))
-
-  // Max Pain
-  let maxPain: number | null = null
-  if (maxPainRes.status === 'fulfilled' && !maxPainRes.value[1]) {
-    maxPain = maxPainRes.value[0]?.data?.max_pain ?? null
-    if (maxPain !== null) {
-      addLog(
-        mkLog('info', 'upstox/max-pain', `Upstox Max Pain Strike=${maxPain}`),
-      )
-      sourceUpdate('upstox/max-pain', 'ok')
-    } else {
-      sourceUpdate('upstox/max-pain', 'error')
-    }
-  } else {
-    sourceUpdate('upstox/max-pain', 'error')
-  }
-
-  // Support and Resistance walls from optionChain Open Interest
-  let supportWall: number | null = null
-  let resistanceWall: number | null = null
-  if (optionChain.length > 0) {
-    let maxPutOi = -1
-    let maxCallOi = -1
-    for (const strike of optionChain) {
-      const putOi = strike.put_options?.market_data?.oi ?? 0
-      const callOi = strike.call_options?.market_data?.oi ?? 0
-      if (putOi > maxPutOi) {
-        maxPutOi = putOi
-        supportWall = strike.strike_price
-      }
-      if (callOi > maxCallOi) {
-        maxCallOi = callOi
-        resistanceWall = strike.strike_price
-      }
-    }
-  }
-
-  // Compute straddle IV from option chain
-  const straddleIv = computeStraddleIV(optionChain, niftyLtp, vix)
-  addLog(
-    mkLog(
-      'debug',
-      'straddle-iv',
-      'ATM IV=' +
-        straddleIv.currentIv +
-        ' vs VIX=' +
-        vix +
-        ' -> ' +
-        (straddleIv.percentAboveAvg !== null
-          ? straddleIv.percentAboveAvg.toFixed(1)
-          : 'null') +
-        '% above avg',
-    ),
-  )
-
-  const adRatio = breadth?.ratio ?? null
-  const proxyValue = computeProxyValuation(niftyLtp, indicators, vix, adRatio)
-  sourceUpdate('synthetic/value', 'ok')
-
-  const niftyPe = { pe: proxyValue.pe, label: proxyValue.label }
-  addLog(
-    mkLog(
-      'info',
-      'synthetic/value',
-      'Computed proxy Nifty PE valuation=' +
-        proxyValue.pe +
-        ' (' +
-        proxyValue.label +
-        ')',
-    ),
-  )
-
-  // Synthetic MMI
-  const mmi = computeMMI(vix, indicators.rsi.value, effectivePcr)
-  addLog(
-    mkLog(
-      'info',
-      'mmi',
-      'Computed proxy MMI score=' +
-        mmi.score +
-        ' (' +
-        mmi.label +
-        ') [vix=' +
-        vix +
-        ' rsi=' +
-        indicators.rsi.value.toFixed(1) +
-        ' pcr=' +
-        effectivePcr.toFixed(3) +
-        ']',
-    ),
-  )
-
-  // News Alerts
-  let newsAlerts: NewsAlert[] = []
-  if (newsRes.status === 'fulfilled' && !newsRes.value[1]) {
-    const rawNews = newsRes.value[0]?.data ?? {}
-    const newsItems = Array.isArray(rawNews)
-      ? rawNews
-      : Object.values(rawNews).flat()
-    try {
-      newsAlerts = classifyNews(newsItems)
-      addLog(
-        mkLog(
-          'info',
-          'upstox/news',
-          `Fetched ${newsItems.length} news articles; classified ${newsAlerts.length} event alerts.`,
-        ),
-      )
-      sourceUpdate('upstox/news', 'ok')
-    } catch (e) {
-      addLog(
-        mkLog('error', 'upstox/news', 'classification failed: ' + String(e)),
-      )
-      sourceUpdate('upstox/news', 'error')
-    }
-  } else {
-    sourceUpdate('upstox/news', 'error')
-  }
-
-  return {
-    mmi: { score: mmi.score, label: mmi.label },
-    advancesDeclines:
-      breadth !== null && breadth.advances !== null
-        ? {
-            advances: breadth.advances,
-            declines: breadth.declines,
-            ratio: breadth.ratio,
-            label: null,
-          }
-        : null,
-    fiiLongShort: fiiLongShort,
-    fiiPositioning: fiiPositioning,
-    pcr:
-      effectivePcr > 0
-        ? {
-            value: parseFloat(effectivePcr.toFixed(3)),
-            zone:
-              effectivePcr >= 1.6
-                ? 'Overbought'
-                : effectivePcr >= 1.0
-                  ? 'Bullish'
-                  : effectivePcr > 0.7
-                    ? 'Neutral'
-                    : 'Bearish',
-          }
-        : null,
-    straddleIv: {
-      elevated:
-        straddleIv.percentAboveAvg !== null && straddleIv.percentAboveAvg > 30,
-      percentAboveAvg: straddleIv.percentAboveAvg,
-    },
-    niftyPe: niftyPe,
-    vix,
-    giftNifty,
-    supportWall,
-    resistanceWall,
-    maxPain,
-    newsAlerts,
-    fetchedAt: new Date().toISOString(),
-  }
-}
-
-// ─── Market fetch ─────────────────────────────────────────────────────────────
-async function fetchMarket(
-  token: string,
-  addLog: (l: BotLog) => void,
-  sourceUpdate: (k: string, s: SourceStatus) => void,
-): Promise<{
-  candles: Candle[]
-  optionChain: OptionData[]
-  v3: V3OrderType
-  breadth: {
-    advances: number
-    declines: number
-    ratio: number
-    total: number
-  } | null
-  globalIndices: GlobalIndexItem[]
-  giftNifty: VrdData['giftNifty']
-}> {
-  addLog(
-    mkLog('debug', 'market', 'fetching candles + breadth + option contracts'),
-  )
-
-  sourceUpdate('candles', 'pending')
-  sourceUpdate('breadth', 'pending')
-  sourceUpdate('option-chain', 'pending')
-  sourceUpdate('global-sentiment', 'pending')
-
-  const [candleRes, breadthRes, contractsRes, globalRes] =
-    await Promise.allSettled([
-      safeFetch<{ data?: { candles?: Candle[] } }>(
-        '/api/market/candles/intraday',
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            token,
-            instrumentKey: 'NSE_INDEX|Nifty 50',
-            interval: '1minute',
-          }),
-        },
-      ),
-      safeFetch<{
-        advances: number
-        declines: number
-        ratio: number
-        total: number
-      }>('/api/market/breadth', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      }),
-      safeFetch<{
-        expiries?: string[]
-      }>('/api/market/option-contracts', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      }),
-      safeFetch<{
-        status: string
-        data?: GlobalIndexItem[]
-        giftNifty: VrdData['giftNifty']
-      }>('/api/market/upstox/global-indices', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      }),
-    ])
-
-  // candles
-  let candles: Candle[] = []
-  if (candleRes.status === 'fulfilled') {
-    const [data, err] = candleRes.value
-    if (err) {
-      addLog(mkLog('error', 'candles', err))
-      sourceUpdate('candles', 'error')
-    } else {
-      candles = data?.data?.candles ?? []
-      addLog(mkLog('info', 'candles', candles.length + ' candles loaded'))
-      sourceUpdate('candles', candles.length > 0 ? 'ok' : 'error')
-    }
-  } else {
-    addLog(mkLog('error', 'candles', 'fetch failed'))
-    sourceUpdate('candles', 'error')
-  }
-
-  // breadth
-  let breadth: {
-    advances: number
-    declines: number
-    ratio: number
-    total: number
-  } | null = null
-  if (breadthRes.status === 'fulfilled' && !breadthRes.value[1]) {
-    breadth = breadthRes.value[0]
-    addLog(
-      mkLog(
-        'info',
-        'breadth',
-        'Nifty 50 A/D loaded: ' +
-          breadth?.advances +
-          '↑ ' +
-          breadth?.declines +
-          '↓ ratio=' +
-          breadth?.ratio,
-      ),
-    )
-    sourceUpdate('breadth', 'ok')
-  } else {
-    const err =
-      breadthRes.status === 'fulfilled'
-        ? (breadthRes.value[1] ?? 'unknown')
-        : 'fetch failed'
-    addLog(mkLog('error', 'breadth', err))
-    sourceUpdate('breadth', 'error')
-  }
-
-  // option chain
-  const contractsData =
-    contractsRes.status === 'fulfilled' && !contractsRes.value[1]
-      ? contractsRes.value[0]
-      : null
-  const contractsErr =
-    contractsRes.status === 'fulfilled' ? contractsRes.value[1] : 'fetch failed'
-  const expiryCandidates = (contractsData?.expiries ?? []).slice(0, 5)
-
-  let optionChain: OptionData[] = []
-  let optionChainError =
-    contractsErr ?? 'No live expiry returned from Upstox option contracts'
-  for (const candidate of expiryCandidates) {
-    const [data, err] = await safeFetch<{ data?: OptionData[] }>(
-      '/api/market/option-chain',
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token, expiryDate: candidate }),
-      },
-    )
-    if (err) {
-      optionChainError = candidate + ': ' + err
-      addLog(
-        mkLog(
-          'warn',
-          'option-chain',
-          'expiry ' + candidate + ' failed: ' + err,
-        ),
-      )
-      continue
-    }
-    const chain = data?.data ?? []
-    if (!chain.length) {
-      optionChainError = candidate + ': empty chain'
-      addLog(
-        mkLog(
-          'warn',
-          'option-chain',
-          'expiry ' + candidate + ' returned empty chain',
-        ),
-      )
-      continue
-    }
-    optionChain = chain
-    addLog(
-      mkLog(
-        'info',
-        'option-chain',
-        optionChain.length + ' strikes loaded (expiry: ' + candidate + ')',
-      ),
-    )
-    sourceUpdate('option-chain', 'ok')
-    break
-  }
-
-  if (!optionChain.length) {
-    addLog(mkLog('error', 'option-chain', optionChainError))
-    sourceUpdate('option-chain', 'error')
-  }
-
-  // Global Sentiment
-  let globalSentiment: ReturnType<typeof evaluateGlobalSentiment> = 'neutral'
-  let globalSentimentFetched = false
-  let globalIndices: GlobalIndexItem[] = []
-  let giftNifty: VrdData['giftNifty'] = null
-  if (globalRes.status === 'fulfilled' && !globalRes.value[1]) {
-    const gData = globalRes.value[0]?.data ?? []
-    globalIndices = gData
-    globalSentiment = evaluateGlobalSentiment(gData)
-    globalSentimentFetched = true
-    giftNifty = globalRes.value[0]?.giftNifty ?? null
-    addLog(
-      mkLog(
-        'info',
-        'global-sentiment',
-        `Global Indices: DJI/NASDAQ/DAX/GIFT rating = ${globalSentiment}`,
-      ),
-    )
-    sourceUpdate('global-sentiment', 'ok')
-  } else {
-    const err =
-      globalRes.status === 'fulfilled'
-        ? (globalRes.value[1] ?? 'unknown')
-        : 'fetch failed'
-    addLog(mkLog('error', 'global-sentiment', err))
-    sourceUpdate('global-sentiment', 'error')
-  }
-
-  // V3
-  let v3: V3OrderType = 'hold'
-  let niftySentiment: NiftySentiment = 'neutral'
-  let pcrZone: ReturnType<typeof evaluatePCR> = 'neutral'
-  let niftySentimentFetched = false
-  let pcrZoneFetched = false
-
-  if (breadth && breadth.advances !== null) {
-    niftySentiment = evaluateNiftySentimentFromAdvanceCount(breadth.advances)
-    niftySentimentFetched = true
-  }
-
-  const totalPut = optionChain.reduce(
-    (sum, item) => sum + item.put_options.market_data.oi,
-    0,
-  )
-  const totalCall = optionChain.reduce(
-    (sum, item) => sum + item.call_options.market_data.oi,
-    0,
-  )
-  if (totalCall > 0) {
-    pcrZone = evaluatePCR(totalPut / totalCall)
-    pcrZoneFetched = true
-  }
-
-  try {
-    if (globalSentimentFetched || niftySentimentFetched || pcrZoneFetched) {
-      v3 = getV3Signal(globalSentiment, niftySentiment, pcrZone)
-      addLog(
-        mkLog(
-          'info',
-          'v3',
-          'signal=' +
-            v3 +
-            ' | global=' +
-            globalSentiment +
-            ' | nifty=' +
-            niftySentiment +
-            ' | pcr=' +
-            pcrZone,
-        ),
-      )
-    }
-  } catch (e) {
-    addLog(mkLog('error', 'v3', 'compute failed: ' + (e as Error).message))
-  }
-
-  return { candles, optionChain, v3, breadth, globalIndices, giftNifty }
-}
-
-function getLotSizeForSymbol(symbol: string): number {
-  const upper = symbol.toUpperCase()
-  if (upper.includes('BANKNIFTY') || upper.includes('NIFTY BANK')) return 15
-  if (upper.includes('FINNIFTY')) return 40
-  if (upper.includes('NIFTY 50') || upper.includes('NIFTY')) return 25
-  return 1
 }
 
 // ─── Hook ──────────────────────────────────────────────────────────────────────
@@ -1333,15 +293,13 @@ export function useStrategyBot(token: string | null) {
       srcUpdates[k] = s
     }
 
-    // Wraps addLog so tick-level logs use the local buffer
-
     log('info', 'tick', `state=${cur.state} trades=${cur.tradesCount}`)
 
     try {
       const config = getStrategyConfig()
       let positionUpdate: Partial<BotStatus> = {}
 
-      // Step 1: fetch candles + option chain + V3 signal in parallel
+      // Step 1: fetch candles + option chain + V3 signal in parallel (from marketService)
       const market = await fetchMarket(token, (e) => tickLogs.push(e), srcUpd)
       const { candles, optionChain, v3, breadth, globalIndices, giftNifty } =
         market
@@ -1381,7 +339,7 @@ export function useStrategyBot(token: string | null) {
       // Step 2: compute indicators first (needed by fetchMarketSentiment for MMI)
       const indicators = computeAllIndicators(candles, optionChain)
 
-      // Step 3: fetch market sentiment (VIX, breadth, NSE PE, FII) using indicator values
+      // Step 3: fetch market sentiment (VIX, breadth, NSE PE, FII) using indicator values (from marketService)
       const vrdData = await fetchMarketSentiment(
         token,
         (e) => tickLogs.push(e),
@@ -1574,7 +532,9 @@ export function useStrategyBot(token: string | null) {
             const executionMode: ExecutionMode = config.executionMode
             const lotSize =
               optionChain.length > 0
-                ? getLotSizeForSymbol(optionChain[0].call_options.instrument_key)
+                ? getLotSizeForSymbol(
+                    optionChain[0].call_options.instrument_key,
+                  )
                 : 25
             let qty =
               finalSignal.positionSize === 'full' ? lotSize * 2 : lotSize
@@ -2023,7 +983,7 @@ export function useStrategyBot(token: string | null) {
       addLogs([...tickLogs, mkLog('error', 'tick', `unhandled: ${msg}`)])
       updateStatus({ error: msg })
     }
-  }, [token, updateStatus, addLog, addLogs])
+  }, [token, updateStatus, addLogs, addLog])
 
   // ── Start / stop ─────────────────────────────────────────────────────────────
   const start = useCallback(() => {
